Repackage automation tests
